Turning the huge formLayout file (and functions) into a bunch of smaller files as react components. Should make things easier to manage
