+++ client/src/features/charts/plugins/functionality/legendBarChart.js
