+++ client/src/pages/chart/Chart.js
@@ -1,4 +1,4 @@
-import ChartMenu from "../../features/charts/sections/ChartMenu";
+import ChartMenu from "../../features/charts/pages/ChartMenu";
 import { ChartProvider } from "../../features/charts/context/chartStore";
 import "./Chart.css";
